elastitest/requirements.txt
Django >= 1.9.6
djangorestframework == 3.3.3
django-extensions >= 1.6.7

